backend/requirements.txt
fastapi
uvicorn[standard]
pillow
numpy
opencv-python-headless
pytesseract
python-dotenv
requests
openfoodfacts
python-multipart
supabase
google-genai


# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
pytest-cov>=4.0.0
